\name{fspval}
\alias{fspval}
\title{fspval}
\usage{
fspval()
}
\description{
X: input dataset (3-mode tensor),
kernel: kernel function,
rho: the parameter associated with the Gaussian kernel,
p: dimension for the low-rank representation,
cl_fun: the clustering algorithm,
rectime: the time record,
toltime: total time record,
lambd: regularization term in the ridge regression,
ndraws: number of draws for the Monte Carlo sampling for approximate the selective p-value.
}
\examples{

}
